Extract theme manager view rendering
